# zimcheck: better and faster redirect loop check