Fixes #161

Depends on openzim/libzim#716

With this implementation deep chains of redirections are not mis-reported as loops. Besides it is faster compared to the old
implementation for the following reasons:

- Redirection info is read from every entry/dirent exactly once; all subsequent processing is with minimal in-memory data required for the task.

- When a standalone loop redirection is performed (for example, `zimcheck -L` with no other option) the auxiliary effient-order-to-by-path-order conversion table is not computed. In this case, in addition to shorter runtime, the memory usage is lower, too.


Benchmark data for `zimcheck -L` [`stackoverflow.com_en_all_2022-05.zim`](http://download.kiwix.org/zim/stack_exchange/stackoverflow.com_en_all_2022-05.zim):

|    |  Run time | Memory usage |
|--------------------------------------------|-------------------|---------------|
| Old implementation (cold start):   |  631 seconds | 1097 MiB |
| Old implementation (warm start): |  500 seconds | 1097 MiB |
| This PR (cold start):                      |  287 seconds | 927 MiB |
| This PR (warm start):                      |  275 seconds | 927 MiB |