@veloman-yunkan 2022-07-27
CI fails because of the dependence on openzim/libzim#716