@veloman-yunkan 2022-08-03
openzim/libzim#716 was merged and participated in the [nightly CI build](https://github.com/kiwix/kiwix-build/actions/runs/2786315512). I reran the CI here today. Macos picked up the change and passed, but Linux failed since it refers to the xenial version of deps, which are no longer being built by the kiwix-build CI.