@veloman-yunkan 2022-08-06
> @veloman-yunkan Two small "problems" are reported by Codefactor.

Not anymore