@veloman-yunkan 2022-08-09
> I wonder if we could do even better on the algorithm.

You try to save runtime by using more memory (`uint32_t` instead of `uint_8`). Sometimes, higher memory usage can result in slower execution. So, though your approach will definitely work, it's not a fact that it will be faster. And I think that memory usage is more important than the runtime - if a program is simply slow you can wait till it completes, whereas if it requires too much memory it simply doesn't run.