@kelson42 2022-08-10
I have no opinion on this specific algorithm question because I don't have the numbers. But, regarding zimcheck, the speed is definitly the priority point to take in consideration. That said, the problem reported by @veloman-yunkan is real: the memory usage should be "reasonable" and should not scale proportionaly to the number of errors/warnings or even size of the ZIM. To conclude here, if we can do significantly faster without using significantly more memory, then this is welcome IMO. Otherwise, probably not.