@mgautierfr 2022-08-10
> That said, the problem reported by @veloman-yunkan is real: the memory usage should be "reasonable" and should not scale proportionally to the number of errors/warnings or even size of the ZIM

My idea will have a memory usage growing proportionally to the number of entry in the ZIM file. (8\*nbEntries+1\*nbRedirectionChains). It is more than @veloman-yunkan solution (5*nbEntries) but it is in the proportion of what we already do in libzim:
- We mmap the urlPtrList (8*nbEntries)
- We mmap the titlePtrList (4*nbEntries)
- For fast iteration, we also create a vector (8*nbEntries) to store the entries order

On top of that, we have to remember that the performance improvement is theoretical for now, we must do some testing to be sure that one solution is more performant than the other.