@veloman-yunkan 2022-08-11
> My idea will have a memory usage growing proportionally to the number of entry in the ZIM file. (8*nbEntries+1*nbRedirectionChains). It is more than @veloman-yunkan solution (5*nbEntries) but it is in the proportion of what we already do in libzim:
> 
>     * We mmap the urlPtrList (8*nbEntries)
> 
>     * We mmap the titlePtrList (4*nbEntries)

`mmap`-ed memory shouldn't be accounted in this analysis as 100% memory consumption since it can be easily discarded by the OS if there is not enough RAM for the real [working set size](https://en.wikipedia.org/wiki/Working_set_size).

>     * For fast iteration, we also create a vector (8*nbEntries) to store the entries order

This one can be halved.

> On top of that, we have to remember that the performance improvement is theoretical for now, we must do some testing to be sure that one solution is more performant than the other.

During redirect loop detection the working set is limited to the data maintained in `RedirectionTable` (i.e. `(8*nbEntries+1*nbRedirectionChains)` vs `(5*nbEntries)`<sup>1</sup>. The size of that data, the size of the CPU data cache, the latency of the memory bus and the access patterns (coming from the redirections) will affect the actual performance. I don't think that spending time on evaluating the performance on multiple ZIM files and diverse hardware is justified in the pursuit  of (some potential) extra speed-up in one of the cheapest checks of `zimcheck` (at the cost of sacrificing the memory usage improvement brought by this PR).

---

<sup>1</sup> the latter, BTW, could be reduced to `(4*nbEntries)` by not maintaining a separate `loopStatus` table and reusing `redirTable` for that purpose, but I didn't like that hack because of the reduced clarity.