@mgautierfr 2022-08-11
I agree that you current algorithm is already a big improvement compared to the previous one. Let's move on with it. We may come back on this when/if we want to improve it further more.

Let's fix the discussion about std::deque/std::vector and we will be good.