@veloman-yunkan 2022-08-12
The Packages CI flows fail since they don't see the change brought by openzim/libzim#716.