@veloman-yunkan 2022-08-12
@kelson42 Should I merge? Or we better resolve the [issue with the Packages builds](https://github.com/openzim/zim-tools/pull/312#issuecomment-1212888345) first?